feat: rename file and add command handler
